RamaEdu Push 09/02

+++ src/main/java/com/g3devs/cliente/juegos/Dado.java
@@ -11,7 +11,7 @@ public class Dado {
 	
 	public int roll() {
 		
-		return null;
+		return (Integer) null;
 	}
 
 }
